policiesABM finished

+++ database/migrations/2024_05_08_231529_create_policies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('policies', function (Blueprint $table) {
             $table->id();
             $table->string('description');
-            $table->float('value');
+            $table->decimal('value', 10, 2);
             $table->timestamps();
         });
     }
